Implement struct only restriction
