+++ src/core-enumeration/Enumeration/Enumeration/Enumeration.cs
@@ -2,7 +2,8 @@
 
 namespace System
 {
-    public static partial class Enumeration<TEnumeration> where TEnumeration : struct
+    public static partial class Enumeration<TEnumeration>
+        where TEnumeration : struct
     {
     }
 }
